# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JJK RPG Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shop/purchase:
    post:
      operationId: purchaseShopItem
      tags: [shop]
      summary: Purchase a shop item with Yen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "200":
          description: Purchase successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResult"
        "400":
          description: Insufficient funds or invalid item

  # ── INVENTORY ─────────────────────────────────────
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: Get player transaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 20
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /transactions/manual:
    post:
      operationId: createManualPurchase
      tags: [transactions]
      summary: Request a manual Yen purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualPurchaseInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualPurchase"
...
components:
  schemas:
...
    ShopItem:
      type: object
      required: [id, name, description, price, type]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        price:
          type: number
        type:
          type: string
        effect:
          type: ["string", "null"]
        image_url:
          type: ["string", "null"]
        use_description:
          type: ["string", "null"]
...
    PurchaseInput:
      type: object
      required: [item_id]
      properties:
        item_id:
          type: number
        quantity:
          type: number
    PurchaseResult:
      type: object
      required: [success, message, new_yen]
      properties:
        success:
          type: boolean
        message:
          type: string
        new_yen:
          type: number
        item:
          $ref: "#/components/schemas/ShopItem"

    # ── INVENTORY ─────────────────────────────────────
...
    Transaction:
      type: object
      required: [id, user_id, type, amount, currency, status, created_at]
      properties:
        id:
          type: number
        user_id:
          type: number
        display_name:
          type: ["string", "null"]
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        yen_amount:
          type: ["integer", "null"]
        elixir_amount:
          type: ["integer", "null"]
        status:
          type: string
        provider:
          type: ["string", "null"]
        provider_tx_id:
          type: ["string", "null"]
        admin_note:
          type: ["string", "null"]
        created_at:
          type: string
    ManualPurchaseInput:
      type: object
      required: [yen_amount]
      properties:
        yen_amount:
          type: number
        contact_note:
          type: ["string", "null"]
    ManualPurchase:
      type: object
      required: [id, user_id, yen_amount, status, created_at]
      properties:
        id:
          type: number
        user_id:
          type: number
        yen_amount:
          type: number
        contact_note:
          type: ["string", "null"]
        status:
          type: string
        admin_note:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
...